backend/pyproject.toml
[project]
name = "plainfigures"
version = "0.1.0"
description = ""
authors = [
    {name = "Jovinder Singh", email = "jovinder@yahoo.com"},
]
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.2",
    "pypdf>=4.3.1",
    "openai>=1.99.6",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "fastapi",
    "langchain>=0.3.19",
    "langchain-core>=0.3.0",
    "langchain-openai",
    "langchain-google-genai",
    "langgraph>=0.2.6",
    "langgraph-sdk>=0.1.57",
    "google-genai",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "numpy>=1.24.0",
    "prophet>=1.1.7",
    "statsmodels>=0.14.5",
    "llama-index>=0.10.0",
    "llama-index-vector-stores-faiss>=0.1.0",
    "llama-index-embeddings-openai>=0.1.0",
    "llama-index-readers-file>=0.5.0",
    "pymupdf>=1.23.0",
    "faiss-cpu>=1.7.4",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-google-vertexai>=2.1.2",
    "google-cloud-aiplatform>=1.114.0",
]

requires-python = ">=3.11,<3.13"

[dependency-groups]
dev = [
    "pytest~=8.3.2",
    "pytest-asyncio~=0.23.7",
    "nest-asyncio>=1.6.0",
    "ipykernel>=6.30.0",
    "easyocr>=1.7.2",
]

[project.optional-dependencies]

jupyter = [
    "jupyter~=1.0.0",
]

lint = [
    "ruff>=0.4.6",
    "mypy~=1.15.0",
    "codespell~=2.2.0",
    "types-pyyaml~=6.0.12.20240917",
    "types-requests~=2.32.0.20240914",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["backend", "frontend"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
